requirements.txt
tensorflow-gpu==2.10.1
scikit-learn==1.3.2
scipy==1.11.4
rdp==0.8
numpy==1.26.1
pillow==10.1.0
pycairo==1.25.1